+++ src/dialog/logdialog.cpp
@@ -116,11 +116,16 @@ void LogDialog::loadSettings()
 {
     OcSettings settings;
     settings.beginGroup("LogWindow");
+    if (settings.contains("geometry")) {
+        restoreGeometry(settings.value("geometry").toByteArray());
+    }
+
+    //remove old settings if they exist
     if (settings.contains("size")) {
-        resize(settings.value("size").toSize());
+        settings.remove("size");
     }
     if (settings.contains("pos")) {
-        move(settings.value("pos").toPoint());
+        settings.remove("pos");
     }
     settings.endGroup();
 }
@@ -130,8 +135,7 @@ void LogDialog::saveSettings()
 
     OcSettings settings;
     settings.beginGroup("LogWindow");
-    settings.setValue("size", size());
-    settings.setValue("pos", pos());
+    settings.setValue("geometry", saveGeometry());
     settings.endGroup();
 }